requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
pyodbc==5.2.0
requests == 2.32.3
python-dotenv == 1.0.1
